// GradesAppFront/src/app/models/alumno-details.interface.ts
// Generated by https://quicktype.io

export interface AlumnoDetailsResponse {
    id:              string;
    nombre:          string;
    apellido:        string;
    fechaNacimiento: string;
    telefono:        string;
    email:           string;
    username:        string;
    password:        string;
}